save scores to scores.json file print scores on input 'scoreboard' from user

+++ index.js
@@ -1,5 +1,7 @@
 const readline = require('readline');
+const fs = require('fs');
 const shuffle = require('./utils/shuffle');
+const createFileIfNotExists = require('./utils/createFileIfNotExists');
 
 let aiChar, userChar, aiMove, userMove;
 
@@ -7,7 +9,6 @@ let aiChar, userChar, aiMove, userMove;
 const init = (board = [], availableMoves = []) => {
 	aiChar = Math.round(Math.random());
 	console.clear();
-	const currentPlayer = 0;
 
 	if (aiChar == 0) {
 		userChar = 1;
@@ -109,31 +110,48 @@ const makeComputerMove = (board, availableMoves) => {
 	board[x][y] = aiMove;
 };
 
+const makeUserMove = (board, availableMoves, move) => {
+	let cords = move.split(' ')[1].trim().split(',');
+	x = parseInt(cords[0], 10);
+	y = parseInt(cords[1], 10);
+	if (x > 2 || y > 2) {
+		console.log('Invalid move. Input must be less than 2');
+		return false;
+	}
+	idx = availableMoves.findIndex((a) => a[0] == x && a[1] == y);
+	if (idx == -1) {
+		console.log('The space is already occupied.');
+		return false;
+	}
+	availableMoves.splice(idx, 1);
+	board[x][y] = userMove;
+	return true;
+};
+
+const printScoreboard = () => {
+	const scores = require('./score.json');
+	console.table(scores);
+};
+
 const processUserMove = (board, availableMoves, move) => {
 	let idx = 0,
 		x = 0,
 		y = 0;
 	if (move.includes('move')) {
-		let cords = move.split(' ')[1].trim().split(',');
-		x = parseInt(cords[0], 10);
-		y = parseInt(cords[1], 10);
-		if (x > 2 || y > 2) {
-			console.log('Invalid move. Input must be less than 2');
-			return false;
-		}
-		idx = availableMoves.findIndex((a) => a[0] == x && a[1] == y);
-		if (idx == -1) {
-			console.log('The space is already occupied.');
-			return false;
-		}
-		availableMoves.splice(idx, 1);
-		board[x][y] = userMove;
-		return true;
+		return makeUserMove(board, availableMoves, move);
+	}
+
+	if (move === 'scoreboard') {
+		printScoreboard();
+		return false;
 	}
 	if (move.includes('quit')) {
 		console.log('Thank you for playing!');
 		process.kill(0);
 	}
+
+	console.log('Invalid input.');
+	return false;
 };
 
 const printBoard = (board) => {
@@ -145,7 +163,7 @@ const play = async (currentPlayer) => {
 	let availableMoves = [];
 	[board, availableMoves] = init(board, availableMoves);
 	let result = '';
-
+	let next = true;
 	while (result === '' && availableMoves.length !== 0) {
 		if (currentPlayer === aiChar) {
 			makeComputerMove(board, availableMoves);
@@ -153,24 +171,39 @@ const play = async (currentPlayer) => {
 		} else if (currentPlayer == userChar) {
 			console.log('It is your turn to make a move.');
 			const move = await getCoordinates();
-			const next = processUserMove(board, availableMoves, move);
+			next = processUserMove(board, availableMoves, move.toLowerCase());
 			if (next) currentPlayer = aiChar;
 		}
-		printBoard(board);
+		if (next) printBoard(board);
 		result = checkWinner(board);
 	}
+
+	const scores = require('./score.json');
 
 	if (result === '' && availableMoves.length === 0) {
 		console.log('The Game is a Draw.');
 	} else if (result === userMove) {
 		console.log('You win!');
+		scores.user += 1;
 	} else if (result === aiMove) {
 		console.log('Computer wins :(');
+		scores.computer += 1;
 	}
+
+	fs.writeFile('./score.json', JSON.stringify(scores), function writeJSON(err) {
+		if (err) return console.log(err);
+	});
 };
 
-const main = () => {
-	play(0);
+const main = async () => {
+	try {
+		const fileName = './score.json';
+		createFileIfNotExists(fileName);
+		play(0);
+	} catch (error) {
+		console.log(error);
+		process.kill(0);
+	}
 };
 
 main();
